Rename guards to app
