+++ modules/firebase/AuthProvider.ts
@@ -7,7 +7,7 @@
 //
 import { useNavigate } from '@tanstack/react-router'
 import { type ReactNode, useEffect } from 'react'
-import { auth } from '@/modules/firebase/guards'
+import { auth } from '@/modules/firebase/app'
 import { routes } from '@/modules/routes'
 import { useAuthUser } from '@/packages/design-system/src/modules/auth/hooks'
