+++ modules/firebase/UserProvider.tsx
@@ -6,7 +6,7 @@
 // SPDX-License-Identifier: MIT
 //
 import { queryOptions, useSuspenseQuery } from '@tanstack/react-query'
-import { getCurrentUser } from '@/modules/firebase/guards'
+import { getCurrentUser } from '@/modules/firebase/app'
 import { getUserInfo } from '@/packages/design-system/src/modules/auth/user'
 
 export const currentUserQueryOptions = () =>
