+++ routes/~__root.tsx
@@ -10,8 +10,8 @@ import { createRootRoute, Outlet, redirect } from '@tanstack/react-router'
 import { NextIntlClientProvider } from 'next-intl'
 import { useLayoutEffect } from 'react'
 import { ReactQueryClientProvider } from '@/app/ReactQueryClientProvider'
+import { auth } from '@/modules/firebase/app'
 import { AuthProvider } from '@/modules/firebase/AuthProvider'
-import { auth } from '@/modules/firebase/guards'
 import messages from '@/modules/messages/translations/en.json'
 import { routes } from '@/modules/routes'
 import { lightTheme } from '@/packages/design-system/src'
