+++ routes/~_dashboard/~users/~index.tsx
@@ -9,7 +9,7 @@ import { UserType } from '@stanfordbdhg/engagehf-models'
 import { createFileRoute, Link } from '@tanstack/react-router'
 import { query, where } from 'firebase/firestore'
 import { UserPlus, Users } from 'lucide-react'
-import { ensureType, getCurrentUser, refs } from '@/modules/firebase/guards'
+import { ensureType, getCurrentUser, refs } from '@/modules/firebase/app'
 import { mapAuthData } from '@/modules/firebase/user'
 import { getDocsData } from '@/modules/firebase/utils'
 import { routes } from '@/modules/routes'
