+++ routes/~_dashboard/~users/~invite.tsx
@@ -9,7 +9,7 @@ import { UserType } from '@stanfordbdhg/engagehf-models'
 import { createFileRoute, useNavigate } from '@tanstack/react-router'
 import { Users } from 'lucide-react'
 import { queryClient } from '@/app/ReactQueryClientProvider'
-import { callables, ensureType } from '@/modules/firebase/guards'
+import { callables, ensureType } from '@/modules/firebase/app'
 import { routes } from '@/modules/routes'
 import { userOrganizationQueryOptions } from '@/modules/user/queries'
 import { PageTitle } from '@/packages/design-system/src/molecules/DashboardLayout'
